backend-AI/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
openai
requests
alembic
PyJWT
bcrypt
langchain
langchain-openai

# For Google OAuth verification
google-auth
